src/third_party/blink/web_tests/fast/html/imports/import-script-expected.txt

CONSOLE WARNING: HTML Imports is deprecated and will be removed in M73, around March 2019. Please use ES modules instead. See https://www.chromestatus.com/features/5144752345317376 for more details.
Test that scripts in imported documents are executed

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS givenDocument is document
PASS givenDocument.currentScript.ownerDocument is leaf.import
PASS givenDocument is document
PASS givenDocument.currentScript.ownerDocument is leaf.import
PASS givenDocument is document
PASS givenDocument.currentScript.ownerDocument is leaf.import
PASS givenDocument is document
PASS givenDocument.currentScript.ownerDocument is leaf.import
PASS givenDocument is document
PASS givenDocument.currentScript.ownerDocument is leaf.import
PASS labels is loadedLabels
PASS successfullyParsed is true

TEST COMPLETE
